+++ tests/Application/lang/en/trans.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'language-key' => 'language-value',
+];